interviewx/ai-models/facial-analysis/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Computer Vision and Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
mtcnn==0.1.1

# Machine Learning and Deep Learning (Updated for Python 3.12 compatibility)
tensorflow==2.17.1
scikit-learn==1.3.2
numpy==1.26.2
scipy==1.11.4

# Additional ML utilities
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2

# Development and Testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Performance
psutil==5.9.5
memory-profiler==0.61.0

# Logging
colorlog==6.7.0

# For model serialization
h5py
protobuf==4.25.5
